+++ CHANGES.md
@@ -1,6 +1,6 @@
 # Release Notes
 
-## Next
+## 2.0.1-preview - 2023-05-17
 
 * PR #69: Fixed dependency shading, reduced the distributable jar size